Split LioCbBuilder out of LioCb
struct LioCb relied on the (incorrect) assumption that a Vec's elements
have a stable location in memory.  That's not true; they can be moved
during Vec::push.  The solution is to use a Vec in the new Builder
struct, but finalize it to a boxed slice (which doesn't support push)
before allowing it to be submitted to the kernel.